changed graphs path to /graphs
